requirement.txt
django-phonenumber-field-7.0.1
phonenumber_field- 1.8.1
corsheaders
djangorestframework
django
Pillow
django-cryptography
virtualenv
djangorestframework-simplejwt
